ft_printf Passes Tripouille tests! My issue was in print_hex where I passed unsigned long when printf itself expects unsigned ints, so I was printing 64 bits instead of 32. I still need to comply with the norm since I'm doing it by typecasting the incorrect type.

+++ Core/ft_printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: lfrench <lfrench@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/13 15:24:31 by lfrench           #+#    #+#             */
-/*   Updated: 2024/04/16 13:05:00 by lfrench          ###   ########.fr       */
+/*   Updated: 2024/04/16 21:58:43 by lfrench          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,9 +54,9 @@ int	ft_print_formatted(char specifier, va_list arg_ptr)
 	else if (specifier == 'u')
 		count += ft_print_unbr(va_arg(arg_ptr, unsigned int));
 	else if (specifier == 'x')
-		count += ft_print_hex(va_arg(arg_ptr, unsigned long), LOWERCASE);
+		count += ft_print_hex(va_arg(arg_ptr, unsigned int), LOWERCASE);
 	else if (specifier == 'X')
-		count += ft_print_hex(va_arg(arg_ptr, unsigned long), UPPERCASE);
+		count += ft_print_hex(va_arg(arg_ptr, unsigned int), UPPERCASE);
 	else
 		count += write(1, &specifier, 1);
 	return (count);
